Module webb_relayer::config

A module for configuring the relayer.

Relayer Configuration Module 🕸️

A module for configuring the relayer.

Overview

The relayer configuration module is responsible for configuring the relayer. Possible configuration include:

port: The port the relayer will listen on. Defaults to 9955

evm: EVM based networks and the configuration. See config/config-6sided-eth-bridge for an example.

substrate: Substrate based networks and the configuration. See config/local-substrate for an example.

Checkout config for useful default configurations for many networks. These config files can be changed to your preferences.

Structs

AnchorBn254PalletConfig

AnchorBn254PalletConfig represents the configuration for the AnchorBn254 pallet.

AnchorContractConfig

AnchorContractOverDKGConfig represents the configuration for the Anchor contract over DKG.

AnchorWithdrawConfig

AnchorWithdrawConfig is the configuration for the Anchor Withdraw.

CommonContractConfig

CommonContractConfig represents the common configuration for contracts.

DKGPalletConfig

DKGPalletConfig represents the configuration for the DKG pallet (dkg-metadata).

DKGProposalHandlerPalletConfig

DKGProposalHandlerPalletConfig represents the configuration for the DKGProposalHandler pallet.

DKGProposalsPalletConfig

DKGProposalsPalletConfig represents the configuration for the DKGProposals pallet.

DkgNodeProposalSigningBackendConfig

DKGNodeSigningBackendConfig represents the configuration for the DKGNode signing backend.

EventsWatcherConfig

EventsWatchConfig is the configuration for the events watch.

EvmChainConfig

EvmChainConfig is the configuration for the EVM based networks.

ExperimentalConfig

ExperimentalConfig is the configuration for the Experimental Options.

FeaturesConfig

FeaturesConfig is the configuration for running relayer with option.

GovernanceBravoDelegateContractConfig

GovernanceBravoDelegateContractConfig represents the configuration for the GovernanceBravoDelegate contract.

LinkedAnchorConfig

LinkedAnchorConfig is the configuration for the linked anchor.

MockedProposalSigningBackendConfig

MockedSigningBackendConfig represents the configuration for the Mocked signing backend.

SignatureBridgeContractConfig

SignatureBridgePalletConfig

SignatureBridgePalletConfig represents the configuration for the SignatureBridge pallet.

SubstrateConfig

SubstrateConfig is the configuration for the Substrate based networks.

SubstrateLinkedAnchorConfig

SubstrateLinkedAnchorConfig is the configuration for the linked anchor of substrate.

TxQueueConfig

TxQueueConfig is the configuration for the TxQueue.

VAnchorBn254PalletConfig

VAnchorBn254PalletConfig represents the configuration for the VAnchorBn254 pallet.

VAnchorContractConfig

VAnchorContractConfig represents the configuration for the VAnchor contract.

WebbRelayerConfig

WebbRelayerConfig is the configuration for the webb relayer.

Enums

Contract

Enumerates the supported contract configurations.

Pallet

Enumerates the supported pallets configurations.

ProposalSigningBackendConfig

Enumerates the supported different signing backends configurations.

SubstrateRuntime

Enumerates the supported Substrate runtimes.

Functions

default_port 🔒

The default port the relayer will listen on. Defaults to 9955.

enable_data_query_default 🔒

Data query access is set to true by default.

enable_leaves_watcher_default 🔒

Leaves watcher is set to true by default.

load

Load the configuration files and

max_blocks_per_step_default 🔒

The maximum events per step is set to 100 by default.

postloading_process 🔒

The postloading_process exists to validate configuration and standardize the format of the configuration

print_progress_interval_default 🔒

The print progress interval is set to 7_000 by default.