Trait webb_relayer::events_watcher::EventHandlerWithRetry

pub trait EventHandlerWithRetry: EventHandler {
    fn handle_event_with_retry<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        store: Arc<Self::Store>, 
        contract: &'life1 Self::Contract, 
        __arg3: (Self::Events, LogMeta), 
        backoff: impl 'async_trait + Backoff + Send + Sync + 'static, 
        metrics: Arc<Metrics>
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
}

An Auxiliary trait to handle events with retry logic.

this trait is automatically implemented for all the event handlers.

Provided Methods

fn handle_event_with_retry<'life0, 'life1, 'async_trait>(
&'life0 self,
store: Arc<Self::Store>,
contract: &'life1 Self::Contract,
__arg3: (Self::Events, LogMeta),
backoff: impl 'async_trait + Backoff + Send + Sync + 'static,
metrics: Arc<Metrics>
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: Sync + 'async_trait,

A method to be called with the event information, it is up to the handler to decide what to do with the event.

If this method returned an error, the handler will be considered as failed and will be retried again, depends on the retry strategy. if you do not care about the retry strategy, use the EventHandler::handle_event method instead.

If this method returns Ok(true), the event will be marked as handled.

Note: this method is automatically implemented for all the event handlers.