Trait webb_relayer::events_watcher::SubstrateBridgeWatcher

pub trait SubstrateBridgeWatcher: SubstrateEventWatcher where
    Self::Store: ProposalStore<Proposal = ()> + QueueStore<TypedTransaction, Key = SledQueueKey> + QueueStore<BridgeCommand, Key = SledQueueKey>, {
    fn handle_cmd<'life0, 'async_trait>(
        &'life0 self, 
        node_name: String, 
        chain_id: U256, 
        store: Arc<Self::Store>, 
        ctx: RelayerContext, 
        client: Arc<Self::Api>, 
        cmd: BridgeCommand
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn run<'life0, 'async_trait>(
        &'life0 self, 
        node_name: String, 
        chain_id: U256, 
        client: Client<Self::RuntimeConfig>, 
        ctx: RelayerContext, 
        store: Arc<Self::Store>
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
}

instead it watches for commands sent from other event watchers or services, it helps decouple the event watchers from the actual action that should be taken depending on the event.

Required Methods

fn handle_cmd<'life0, 'async_trait>(
&'life0 self,
node_name: String,
chain_id: U256,
store: Arc<Self::Store>,
ctx: RelayerContext,
client: Arc<Self::Api>,
cmd: BridgeCommand
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Provided Methods

fn run<'life0, 'async_trait>(
&'life0 self,
node_name: String,
chain_id: U256,
client: Client<Self::RuntimeConfig>,
ctx: RelayerContext,
store: Arc<Self::Store>
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Returns a task that should be running in the background that will watch events