Trait webb_relayer::events_watcher::SubstrateEventWatcher

pub trait SubstrateEventWatcher {
    type RuntimeConfig: Config + Send + Sync + 'static;
    type Api: From<Client<Self::RuntimeConfig>> + Send + Sync;
    type Event: Decode + Send + Sync + 'static;
    type FilteredEvent: Event + Send + Sync + 'static;
    type Store: HistoryStore;

    const TAG: &'static str;

    fn handle_event<'life0, 'async_trait>(
        &'life0 self, 
        store: Arc<Self::Store>, 
        api: Arc<Self::Api>, 
        __arg3: (Self::FilteredEvent, <<Self as SubstrateEventWatcher>::RuntimeConfig as Config>::BlockNumber)
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn run<'life0, 'async_trait>(
        &'life0 self, 
        node_name: String, 
        chain_id: U256, 
        client: Client<Self::RuntimeConfig>, 
        store: Arc<Self::Store>
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
}

Represents a Substrate event watcher.

Required Associated Types

type RuntimeConfig: Config + Send + Sync + 'static

The Config of this Runtime, mostly it will be [subxt::DefaultConfig]

type Api: From<Client<Self::RuntimeConfig>> + Send + Sync

The Runtime API.

type Event: Decode + Send + Sync + 'static

All types of events that are supported by this Runtime. Usually it will be [my_runtime::api::Event] which is an enum of all events.

type FilteredEvent: Event + Send + Sync + 'static

The kind of event that this watcher is watching.

type Store: HistoryStore

Required Associated Constants

const TAG: &'static str

Required Methods

fn handle_event<'life0, 'async_trait>(
&'life0 self,
store: Arc<Self::Store>,
api: Arc<Self::Api>,
__arg3: (Self::FilteredEvent, <<Self as SubstrateEventWatcher>::RuntimeConfig as Config>::BlockNumber)
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Provided Methods

fn run<'life0, 'async_trait>(
&'life0 self,
node_name: String,
chain_id: U256,
client: Client<Self::RuntimeConfig>,
store: Arc<Self::Store>
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Returns a task that should be running in the background that will watch events