Function webb_relayer::handler::handle_text

pub async fn handle_text<TX>(
    ctx: &RelayerContext, 
    v: &str, 
    tx: &mut TX
) -> Result<()> where
    TX: Sink<Message> + Unpin,
    TX::Error: Error + Send + Sync + 'static, 

Sets up a websocket channels for message sending.

This is primarily used for transaction relaying. The intention is that a user will send formatted relay requests to the relayer using the websocket. The command will be extracted and sent to handle_cmd if successfully deserialized.

Returns Ok(()) on success

Arguments

ctx - RelayContext reference that holds the configuration

v - The text (usually in a JSON form) message to be handled.

tx - A mutable Trait implementation of the warp::ws::Sender trait

Examples

let _ = handle_text(ctx, text, &mut tx).await?;;