Module webb_relayer::service

A module for starting long-running tasks for event watching.

Relayer Service Module 🕸️

A module for starting long-running tasks for event watching.

Overview

Services are tasks which the relayer constantly runs throughout its lifetime. Services handle keeping up to date with the configured chains.

Enums

ProposalSigningBackendSelector 🔒

Functions

ignite

Starts all background services for all chains configured in the config file.

make_proposal_signing_backend 🔒

make_substrate_proposal_signing_backend 🔒

start_dkg_pallet_watcher 🔒

Starts the event watcher for DKG pallet events watcher.

start_dkg_proposal_handler 🔒

Starts the event watcher for DKG proposal handler events.

start_evm_anchor_events_watcher 🔒

Starts the event watcher for EVM Anchor events.

start_evm_vanchor_events_watcher 🔒

Starts the event watcher for EVM VAnchor events.

start_signature_bridge_events_watcher 🔒

Starts the event watcher for Signature Bridge contract.

start_substrate_anchor_event_watcher 🔒

Starts the event watcher for Substrate anchor events.

start_substrate_signature_bridge_events_watcher 🔒

Starts the event watcher for Signature Bridge Pallet.

start_substrate_vanchor_event_watcher 🔒

Starts the event watcher for Substrate vanchor events.

start_tx_queue 🔒

Starts the transaction queue task

Type Definitions

Client 🔒

Type alias for providers

DkgClient 🔒

Type alias for the DKG DefaultConfig

DkgRuntime 🔒

Type alias for the DKG RuntimeApi

Store 🔒

Type alias for Sled-based database store

WebbProtocolClient 🔒

Type alias for the WebbProtocol DefaultConfig

WebbProtocolRuntime 🔒

Type alias for the WebbProtocol RuntimeApi