Trait webb_relayer::store::EventHashStore

pub trait EventHashStore: Send + Sync + Clone {
    fn store_event(&self, event: &[u8]) -> Result<()>;
    fn contains_event(&self, event: &[u8]) -> Result<bool>;
    fn delete_event(&self, event: &[u8]) -> Result<()>;
}

A Simple Event Store, that does not store the events, instead it store the hash of the event as the key and the value is just empty bytes.

This is mainly useful to mark the event as processed.

Required Methods

fn store_event(&self, event: &[u8]) -> Result<()>

Store the event in the store. the key is the hash of the event.

fn contains_event(&self, event: &[u8]) -> Result<bool>

Check if the event is stored in the store. the key is the hash of the event.

fn delete_event(&self, event: &[u8]) -> Result<()>

Delete the event from the store. the key is the hash of the event.