Trait webb_relayer::store::HistoryStore

pub trait HistoryStore: Clone + Send + Sync {
    fn set_last_block_number<K: Into<HistoryStoreKey> + Debug>(
        &self, 
        key: K, 
        block_number: U64
    ) -> Result<U64>;
    fn get_last_block_number<K: Into<HistoryStoreKey> + Debug>(
        &self, 
        key: K, 
        default_block_number: U64
    ) -> Result<U64>;

    fn get_last_block_number_or_default<K: Into<HistoryStoreKey> + Debug>(
        &self, 
        key: K
    ) -> Result<U64> { ... }
}

HistoryStore is a simple trait for storing and retrieving history of block numbers.

Required Methods

fn set_last_block_number<K: Into<HistoryStoreKey> + Debug>(
&self,
key: K,
block_number: U64
) -> Result<U64>

Sets the new block number for that contract in the cache and returns the old one.

fn get_last_block_number<K: Into<HistoryStoreKey> + Debug>(
&self,
key: K,
default_block_number: U64
) -> Result<U64>

Get the last block number for that contract. if not found, returns the default_block_number.

Provided Methods

fn get_last_block_number_or_default<K: Into<HistoryStoreKey> + Debug>(
&self,
key: K
) -> Result<U64>

an easy way to call the get_last_block_number where the default block number is 1.