Trait webb_relayer::store::LeafCacheStore

pub trait LeafCacheStore: HistoryStore {
    type Output: IntoIterator<Item = H256>;

    fn get_leaves<K: Into<HistoryStoreKey> + Debug>(
        &self, 
        key: K
    ) -> Result<Self::Output>;
    fn insert_leaves<K: Into<HistoryStoreKey> + Debug>(
        &self, 
        key: K, 
        leaves: &[(u32, H256)]
    ) -> Result<()>;
    fn get_last_deposit_block_number<K: Into<HistoryStoreKey> + Debug>(
        &self, 
        key: K
    ) -> Result<U64>;
    fn insert_last_deposit_block_number<K: Into<HistoryStoreKey> + Debug>(
        &self, 
        key: K, 
        block_number: U64
    ) -> Result<U64>;
}

A Leaf Cache Store is a simple trait that would help in getting the leaves and insert them with a simple API.

Required Associated Types

type Output: IntoIterator<Item = H256>

Required Methods

fn get_leaves<K: Into<HistoryStoreKey> + Debug>(
&self,
key: K
) -> Result<Self::Output>

fn insert_leaves<K: Into<HistoryStoreKey> + Debug>(
&self,
key: K,
leaves: &[(u32, H256)]
) -> Result<()>

fn get_last_deposit_block_number<K: Into<HistoryStoreKey> + Debug>(
&self,
key: K
) -> Result<U64>

fn insert_last_deposit_block_number<K: Into<HistoryStoreKey> + Debug>(
&self,
key: K,
block_number: U64
) -> Result<U64>