Function webb_relayer::tx_relay::substrate::handle_substrate_tx

pub async fn handle_substrate_tx<E: Debug + Decode + HasModuleError, Evs: Decode>(
    event_stream: TransactionProgress<'_, DefaultConfig, E, Evs>, 
    stream: Sender<CommandResponse>
)

Handles a submitted Substrate transaction by processing its TransactionProgress.

The TransactionProgress is a subscription to a transaction’s progress. This method is intended to be used in a variety of places for all kinds of submitted Substrate transactions.