Module webb_relayer::service

A module for starting long-running tasks for event watching.

Relayer Service Module 🕸️

A module for starting long-running tasks for event watching.

Overview

Services are tasks which the relayer constantly runs throughout its lifetime. Services handle keeping up to date with the configured chains.

Enums

ProposalSigningBackendSelector

Proposal signing backend config

Functions

build_web_services

Sets up the web socket server for the relayer, routing (endpoint queries / requests mapped to handled code) and instantiates the database store. Allows clients to interact with the relayer.

ignite

Starts all background services for all chains configured in the config file.

make_proposal_signing_backend

utility to configure proposal signing backend

make_substrate_proposal_signing_backend

utility to configure proposal signing backend for substrate chain

start_dkg_pallet_watcher

Starts the event watcher for DKG pallet events watcher.

start_dkg_proposal_handler

Starts the event watcher for DKG proposal handler events.

start_dkg_substrate_tx_queue

Starts the transaction queue task for dkg-substrate extrinsics

start_evm_open_vanchor_events_watcher

Starts the event watcher for EVM VAnchor events.

start_protocol_substrate_tx_queue

Starts the transaction queue task for protocol-substrate extrinsic

start_signature_bridge_events_watcher

Starts the event watcher for Signature Bridge contract.

start_substrate_signature_bridge_events_watcher

Starts the event watcher for Signature Bridge Pallet.

start_substrate_vanchor_event_watcher

Starts the event watcher for Substrate vanchor events.

start_tx_queue

Starts the transaction queue task

Type Definitions

Client

Type alias for providers

DkgClient

Type alias for the DKG DefaultConfig

Store

Type alias for Sled-based database store

WebbProtocolClient

Type alias for the WebbProtocol DefaultConfig