Module webb_relayer::service

A module for starting long-running tasks for event watching.

Relayer Service Module 🕸️

A module for starting long-running tasks for event watching.

Overview

Services are tasks which the relayer constantly runs throughout its lifetime. Services handle keeping up to date with the configured chains.

Functions

build_web_services

Sets up the web socket server for the relayer, routing (endpoint queries / requests mapped to handled code) and instantiates the database store. Allows clients to interact with the relayer.

ignite

Starts all background services for all chains configured in the config file.

Type Definitions

Client

Type alias for providers

Store

Type alias for Sled-based database store